bugfix to IJ maximumfinder
Maximumfinder was being created as a global static object that was reused with each additional run.  This meant that new images with different dimensions were being assumed to have the same dimensions as the first image run, resulting in issues.  This has now been fixed by moving the MaximumFinder into a local object.